aact-808:  When migrations run, it can only know about the proj schema, otherwise it will create a schema_migration table someplace else.  We need all project migrations to be logged together, and separate from the Core AACT migrations.
We are now populating the AACT Staging Database with Project data, rather than populating the Public DB directly.  This is in coordination with Core AACT - which will include the Project schemas when it does a dump/restore to the public DB.  We are doing this because: A) we want to run sanity checks on it before making it public, B) the project data loads were very slow on the server when it was populating the public DB over the network, C) we need to refactor things so that having views that cross schemas will not break the load.