aact-810:  Use variables to define super user role and db name.  We were going to use a specific postgres user (proj) to manage aact-proj, but think it might be best to use the same postgres user that manages the rest of aact (ctti).  Resisted doing this because we need to set the search_path for managing projects, and I don’t want that to mess up regular aact loads, but we can set the search path per database, so that shouldn’t be a problem.
Set the path when we run rake db:create because db:migrate is going to need to know the schema in which to store the schema_migrations table.  This schema should be ‘public’, which is why it needs to be the first schema in the list when setting the search_path.